Create LocalLogger which just calls print(), and cut @logEntryExit over to use it.
@logEntryExit is a little magic, since before this it wasn't tied to any sort of provider.
Now, it looks on the object it was called on, and looks for the logger object.
If it's not found, it asserts.

This lets any object - even if it's not a provider - use the attribute, even if it
isn't strictly a provider.

An example of this is the MySQLDatabase. Although in that case, we cut it over to look more
like a Provider, have it inherit from INeedsLoggingProvider (but not BaseProvider) and
use an _update_config method on DatabaseProvider.