Fix logging categories
.bind_category was being called on the same instance shared among every Provider.
Therefore the last provider to initialize set the category for every other provider.
We copy the logging object and provider an overridden log function with category defined.

We have to update the test because we are not comparing identical objects anymore,
so we just make sure that it was populated.

Although, since we are duplicating the logging object we could just keep the whole
bind_category idea instead of using a partial, but I think the partial is better?