Refactor JOBSTATUS into two separate parts: Status and Outcome
Status is what updatebot needs to do for it next.
Outcome is how the job completed.

Outcome is really needed more for testing purposes than anything else
because in theory we won't care how it ended as long as it ended but
that's a short-sighted view because in the future we will definetly
want to display that information on a dashboard.