Add the test for testExistingJobUnclassifiedFailuresNeedingRetriggers
Along the way we needed to remove an assertion that was checking for only one of a type
of job. That assertion would probably (?) be true for the first run through, but definetly
not after retriggers. And it's of dubious value on the first run through (it's just asserting
something I expect to be true, not something I rely on (I think)), so let's just nix it.