Port testExistingJobClassifiedFailures
This was done by copying the unclassified failure files and then moving all the failures
from the needsInvestigation section of push health ro knownIssues of push health

I don't think we have any tests that exercise the taskcluster classified failure scenario.
Hopefully this appears in code coverage and we can note it for a future test.