Fix LibraryProvider and test
I noticed that we used assertTrue when we should have used assertEqual,
causing everything the test to pass as long as the value would be cast
to True. The only time this wouldn't happen was would be when 'Enabled'
was False. Unfortunately, we weren't populating the Enabled property
correctly, and had a copy/paste error there!