Refactor the commit alert test setup to support dynamic expected commits
Previously we calculated the expected number of commits based on the
difference between the revision for the library and the tip.

But that calculation can be incorrect because we may have created a job
about one of the commits - but not updated the library's manifest - and
that commit won't be included in a subsequent job.

Way back when I noted that I could have just hardcoded the number of
commits we expected to see.  Now that the expected number of commits has
become a lambda because it can change over the lifetime of the test,
I do go back to the old tests and just hardcode them as a lambda:1 or
whatever.

For the places where it is dynamic, I have lifted the old behavior into
a global variable and a generic function that can be populated with the
correct functions to call via functools.partial