Add commit alert tests for a branch
This also adds our first test that encompasses two upstream commits.
During development of the test, only one commit was actually acknowledged,
but the test still passed.  The bug was fixed, but it illustrated the need
for an additional check in the test to ensure we had seen the expected
number of commits.

While I could hardcode the expected number of commits, I opted to determine
it programmatically by finding the distance between the commits in an array.

I also tweaked the library provider to be a little more friendly.