Add a test that creates two alerts (because upstream has changed after the first)
This required a somewhat complicated trick.  We need a way to have the moz.yaml
change its revision based on how the test progresses. But it isn't very conducive
to that in the current architecture. So instead of giving it a revision, we give
it a callback (a lambda) that it will call whenever it needs to return the current
revision.

That callback's return value depends on a call_counter that the test will advance
manually.  To ensure the lambda would behave as we expected it to, a quick
lambda_capture test was written.

An implication of this change is that expected_commits_seen is no longer a
constant but must become a lambda itself so it can call the callbacks also.

Finally, all the previous tests must change their constant revisions to a lambda
that returns that (same) constant revision.

Nestled in these changes is the same notion of a lambda used for a repo_override
so we can specify test-repo-a.bundle or test-repo-b.bundle. But we don't use it
yet.