Create a test repository bundle we can use for commit alert tests
This test repo contains a series of commits that we'll use for testing
commit alerts. In addition to the repo bundle, we have separate repo
bundles which contain only up to the specified commit.  This is needed
so we can clone a repo with commit A, and then later clone the 'same'
repo again, but it's been updated to have commits A -> B.
(That would be done by cloning repo-a.bundle and then repo-b.bundle)