Change how we process the results for awaiting retrigger results
Previously we assumed that if we got to the awaiting retrigger results,
we had to have unclassified failures. In #171 we discovered that is not
necessarily the case because we might have had failures change status.

So abstract the handling of Classified vs Unclassified Failures into a
function (which also includes checking for build failures when we process
retrigger results, but this is harmless I think) that is called after both
second_platform_results and retrigger_results