Big refactoring to how we lie about Bug IDs to Updatebot
Previously we had a single Bug Id (hardcoded as 50) and to accomodate
the fact that the next bug we file should have a larger bug Id, we had
a lambda function that would provide an offset to add to that single
Bug ID.

For a simple test we would start with a Bug ID of 50, and define a
lambda of `lambda : call_counter`. We would call updatebot once,
increment call_counter, and call it again, and the second time we ran
it we would take the offset of 1 and add it to 50 to get a bug ID of 51.

For a more complicated test, one where we ran updatebot once (and filed
a bug), ran it a second time (and did not file a bug), and then a third
time (and filed a bug) we would do a function like this:

def f():
    if call_counter > 1:
       return 1
    return 0

--------------

This design had a big limitation that manifested in a few ways. We use
Bug IDs for (at least) three things:
 - Figuring out what the Bug ID of the bug we're about to file is
 - Figuring out the Bug IDs of the bugs we've previously filed
 - Figuring out which of the prior bugs we've filed are still open

As we've added more logic that works with the latter two in different
ways (behaving differently based on whether we have prior filed bugs
and whether they are open) we need to exercise that logic by returning
different values for thse three situations, which we previously couldn't
do.

With this commit, we handle the first with a new lambda
'_get_filed_bug_id_func', and the latter two with a new lambda
'filed_bug_ids_func(bool only_open)'.

Like our other lambdas, they use a global call_counter incrementing
integer to determine what value(s) to return when.

--------------

With this new refactoring, we can also improve our assertions in the
file_bug method to ensure that we are properly referencing prior bugs
in both the Depends On field, and the Description text.

We can also remove a parameter from _check_job which is no longer needed.