Switch from specifying a SHA hash to controlling the commit list output from git
We make this a lambda because we need to be able to update the SHA hash so
we can write tests where more than one job will run.

Rather than making a lambda that just returns a hash, we also have to control
the result of the commit list from git - this is because inside
SCMProvider.check_for_updates we perform different things based on whether or
not e.g. the prior bug covered some of the upstream commits we are seeing, but
not all.

For the current tests, however, we are only doing one job in the test, so the
lambda can simply be a constant string.