Test structure improvements for phab revisions/abandoning
With this commit we will create different phabricator revisions
that correspond to the bug numbers we use.

We also add the ability to dtect when a revision is abandoned,
and require the test author to explicitly check that abandoning
is expected.  We add that bit to all prior tests that need it.