Handle Relinquishing a bug
Relinquishing means that Updatebot will no longer do anything
with the bug. Presently this is only used for Vendoring jobs.

This is because Vendoring jobs may be closed by Updatebot if a
more recent job is created. We don't want someone to re-open a
bug in Bugzilla, and then Updatebot come along and re-close it
on them.

So after Updatebot closes a bug it marks the job as Relinquished
so it will no longer close it, or even reference it in the future.

(I don't think not-referencing it will be a problem, but maybe
that will change in the future.)

This is not needed for Commit Alert jobs, because while Commit
Alerts may be updated to mark future FF versions as affected,
it won't close a bug, nor will it re-mark a FF version as
'affected' if, for example, someone marks it as 'wontfix'.