Change SCMProvider logic to use merge-base. Fixes #206
More details are in #206, but the gist is that our current
code assumes a linear history between the commit in-tree
and the one we want to update to.  Thsi may not be the case
if e.g. in-tree we are on branch A with commit M, and
M is only on A and not an ancestor of HEAD.