Add in error handling on virtually all actions during vendoring
This should avoid us getting stuck in infinite loops unless e.g.
we can't save the status to the database.

I'm not in love with the way this makes the code change, it now
looks like C, with 4 lines of error handling for every function
call.  Not sure how to address it though...