Improve behavior on holly
If you're committing and testing on Holly, hg strip()
will delete your commits to holly that aren't in m-c.
So the second library you process will not have your commits.
As we add more libraries to Updatebot this is a bigger deal.
Instead of using strip there, just update to the original commit.