Allow the SCM Provider to clone the repository at different times
Previously there was only one flow that would clone the repository
for the SCM Provider, and it would persist only for the lifetime of
the function.

Now we can call the function twice but only clone it once and it
will persist and can be used by the other call.