Fix a bug that came in from the spurious update feature
Previously we passed in the bug id to create_job() which would
associate it with the returned job object.  We stopped doing that
so created_job did not have a bugzilla_id attached to it.  When you
tried to comment on a bug from it, it was zero and Bugzilla does not
like that.