Fix a test oversight
While developing a new test, it did not fail even
though no job was created. I realized thast the check_job
function was not asserting that it checked anything, only
that if it found something to check, that it was correct.

This adds a check that it did find one, and only one, thing
to check.  Fortunately all tests were working and adding this
check did not cause any failures.