Switch from using a dict to an OrderedDict for test commands
We relied on undefined behavior by not being prefix-free when
defining command lookups.  That is to say you could have two
commands './mach vendor foo' and './mach vendor' and the only
thing that routed './mach vendor foo banana' to the first one
was that it was found first when iterating through the dict
keys.

But that ordering is undefined behavior and it finally caught
up with us and started causing intermittent failures.  So
switch it to an OrderedDict, so it is defined behavior.