Cut over testSecondJobReferencesFirst
Add assertions that both jobs will succeed, now that they will.

In this one the treeherder functions [health/jobs]_calls would be
confusing if we tried to do the 'tell it things are still running the
first time' pattern we usually do.  It's doable but it's not
necessary, that pattern is well tested.  So we make the second
job succeed immediately.