Do not always try to mark a bugzilla bug as affected for a new firefox version
A Spurious Update outcome does _not_ file a bugzilla bug, so there is
nothing to update.  I think that the rest of the outcomes should all
have a bugzilla bug (I'm thinking of the many failure outcomes) but in
case not - raise a more useful Exception than the current one which
will help diagnose future issues.

Also change the function to a string of if/elif because that's more
logical.