Change how we process the results in vendoring
> if results['to_investigate'] and comment_lines

What is this garabage, it makes no sense.  Switch to
a more logical determination of which path to go down.

Also change each path to its own function. They previously
shared a function because they both wound up using the last
`self.dbProvider.update_job_status(existing_job, JOBSTATUS.DONE)`
line but that's just confusing.

In the new functions, if the phab submit fails, it could update
the job status twice, but that should be fine.