Change how we determine what jobs to retrigger
Conceptually make this function a lot simpler by grouping
the results in two ways: by their test (which comes from push health)
and from the jobname.  (The same test may run on different jobs
for different platforms.)

We retrigger a job if
(1) the job ran a test, and the test failed, and there were
    less than 2 runs (fail or pass) of that test, and the failure
    wasn't classified as NotYourFault
(2) the job isn't a built or lint job, and the job wasn't classified
    as NotYourFault, and the job failed, there were less than 2
    runs (fail or pass) of that job, and the job wasn't seen in the
    Push Health results

The last clause "the job wasn't seen in the push health results"
is load bearing.  It is what stops us from e.g.

  a. re-running a job with only one failure, because that failure
     was part of a test failure that only occured on 1 of 4 platforms.
     Could it be a legit failure? Yeah, it could, but it might also
     be an intermittent. I haven't done the analysis to see which is
     more likely, but we don't want to be wasteful with resources
     and this would cause us to retrigger for virtually every
     intermittent
  b. re-running a job that Taskcluster thinks is New Failure or an
     Intermittent, but Push Health classified it as NotYourFault