Introduce a soft timeout
This introduces a "soft_timeout" config option, which specifies after
which ammount of seconds updatebot should abort its run. The check for
this happens after each task. In some cases this can prevent the
external taskcluster timeout from killing updatebot while it is working
on a task, resulting in a unexpected databse state.