Fix handling unexpected_create statuses
In 52da7211c9f812520cfb1f59b9fd8793df74cd46 we will relinquish a job
we find with an unexpected created status.

However, we neglected to relinquish the job _prior_ to this unexpected
job, which caused us to break our invariant that if there is a non-relinquished
job, it must be the most recent one.

Now, when handling an unexpected created status job we relinquish it
and the prior job.

Doing this required us to tack on a linked-list style reference to the prior job.

Fixes #368