Add @retry to two more places
Push Health, cause we just observed a failure there; and job details cause
that seems like it could fail.

I always have a vague recollection that the order of decorators matters
so I'm going to fix on instance and then put in a script to double check
things.  Although truthfully I'm not 100% sure it's necessary.