Shrink the size of the argument to SCMProvider.check_for_update
We don't need the whole job object, just its version is sufficient.

This is an ugly fix for a recursion overflow in the memoize
implementation where we pickle the argument.

A better fix would be to fix memoize.
Or revisit the need to have a linked list of every single job
for a library.

Fixes #403